pyproject.toml
[project]
name = "cortiscan"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "onnx>=1.18.0",
    "onnxruntime>=1.22.1",
    "pillow>=11.3.0",
]
